Port some crate internal only types.
