+++ src/firesatimage.rs
@@ -1,3 +1,99 @@
+use crate::pixel::Pixel;
+
+/**
+ * Handle to a dataset for the Fire Detection Characteristics and some metadata.
+ */
+pub(crate) struct SatFireImage {
+    /// Image width in pixels
+    xlen: usize,
+    /// Image height in pixels
+    ylen: usize,
+    /// All the information needed for transforming from row and column numbers to coordinates.
+    tran: CoordTransform,
+    // In memory buffer if this is from a zip file
+    // unsigned char *memory;
+    // Size of the buffer;
+    // size_t memory_size;
+    // Handle to the NetCDF file
+    // int nc_file_id;
+    /// Orignial file name the dataset was loaded from.
+    fname: String,
+}
+
+/**
+ * Represents all the data associated with a single pixel in which the satellite has detected
+ * a fire.
+ */
+pub(crate) struct FirePoint {
+    /// The polygon describing the scanned area.
+    pixel: Pixel,
+    /// The x-coordinate (column number, often indexed as 'i') in the grid.
+    x: isize,
+    /// The y-coordinate (row number, often indexed as 'j') in the grid.
+    y: isize,
+}
+
+/// Projection information required to convert from row/column number to scan angles and lat-lon.
+#[allow(non_snake_case)]
+struct CoordTransform {
+    /// Scale factor for the column for converting indexes to scan angle coords.
+    xscale: f64,
+    /// Offset for the column for converting indexes to scan angle coords
+    xoffset: f64,
+    /// Scale factor for the row for converting indexes to scan angle coords
+    yscale: f64,
+    /// Offset for the  row for converting indexes to scan angle coords
+    yoffset: f64,
+    /// Radius of the Earth at the equator in meters.
+    req: f64,
+    /// Radius of the Earth at the poles in meters.
+    rpol: f64,
+    /// Height of the satellite above the equator in meters.
+    H: f64,
+    /// Longitude of the nadir point in degrees.
+    lon0: f64,
+}
+
+/*
+
+/**
+ * \brief Open a file containing GOES-R/S Fire Detection Characteristics.
+ *
+ * \param fname the path to the file name to open.
+ * \param tgt the [FireDataSet] structure to initialize.
+ *
+ * \returns false if there is an error opening the data.
+ */
+bool fire_sat_image_open(char const *fname, struct SatFireImage *tgt);
+
+/**
+ * \brief Close the file and clear the pointers associated with  with this dataset.
+ *
+ * \param dataset is the structure to close/finalize.
+ */
+void fire_sat_image_close(struct SatFireImage *dataset);
+
+/**
+ * \brief Extract pixels/points from the image that have non-zero fire power.
+ *
+ * \returns GArray * of struct FirePoint objects.
+ */
+GArray *fire_sat_image_extract_fire_points(struct SatFireImage const *fdata);
+
+/** Add a FirePoint to this Cluster. */
+void satfire_cluster_add_fire_point(struct SFCluster *cluster, struct FirePoint *fire_point);
+
+/** Steal the pixels from this row.
+ *
+ * This is very unsafe because it leaves the struct in an invalid state with a \c NULL pointer
+ * in the row->pixels. Use carefully and finalize/destroy the object soon after calling this
+ * function.
+ *
+ * An alternative would be to call satfire_pixel_list_copy() and get a deep copy of the
+ * \ref SFPixelList. But in some cases that would be a wasteful copy if we are done using the
+ * \ref SFClusterRow now.
+ */
+struct SFPixelList *satfire_cluster_db_satfire_cluster_row_steal_pixels(struct SFClusterRow *row);
 #include <assert.h>
 #include <libgen.h>
 #include <netcdf.h>
@@ -487,3 +583,4 @@ ERR_RETURN:
 
     return 0;
 }
+*/
